Pull views into independent files
