+++ config.ini.example
@@ -6,6 +6,7 @@ PendingRewardChannel =
 Tier1RoleID = 1036801694564102174
 Tier2RoleID = 1036807522457231384
 Tier3RoleID = 1036807951484203099
+GuildID = 82969926125490176
 
 [MySQL]
 Username =
